+++ CHANGELOG.md
@@ -33,7 +33,8 @@
     
 #### Fixed:
 
-- Correctly handle Dropbox server errors.
+- Correctly handle additional error types: internal Dropbox server error, insufficient
+  space on local drive, file name too long for local file system and out-of-memory error.
 - Set the log level for the systemd journal according to user settings instead of using 
   logging.DEBUG.
 - When invoking `maestral restart`, run checks for Dropbox folder location and link status.